Fixed warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@51246 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objmgr/util/sequence.cpp
@@ -3229,7 +3229,7 @@ CRef<CBioseq> CSeqTranslator::TranslateToProtein(const CSeq_feat& cds,
         }
     }
 
-    int prot_len = 0;
+    size_t prot_len = 0;
     ITERATE (CDelta_ext::Tdata, seg_it, prot->SetInst().SetExt().SetDelta().Set()) {
         prot_len += (*seg_it)->GetLiteral().GetLength();
     }
